requirements.txt
Flask
tensorflow-cpu
librosa
gunicorn